Backend/requirements.txt
Flask==2.2.5
knock-subdomains==7.0.2
flask-cors==3.0.10
python-dotenv==0.19.0
flask-jwt-extended==4.3.1
bcrypt==4.0.1
dnspython >= 2.4.2
requests>=2.31.0
python-owasp-zap-v2.4
python-nmap==0.7.1
pymongo==4.6.1
flask-mail
celery==5.3.1
redis==5.0.1
pytz>=2021.1
cohere
